feat(retrieval): add v2 style endpoints
relates #7
